Update README.md
Alembic is a genome finishing program that makes additional joins on pre-assembly contigs.  It calculates gap sizes and overlaps based on one or more reference genomes concurrently.  It uses an iterative pairing approach which does not make assumptions that the order and orientation of genes and non-coding sequence of related species are identical.  This approach yields a more accurate assembly than a simple alignment of one species to another.